+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Added the `renderStr` and `renderStr'` functions for efficient rendering
   to `String` and `ShowS`.
 
+* Added the `parserBs` that can consume strict `ByteString` streams.
+
 ## Modern URI 0.0.1.0
 
 * Initial release.